Move firebaseConfig to src
