+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
-import './auth/firebase'
+import './firebase'
 
 
 ReactDOM.render(
